% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{allTrees2}
\alias{allTrees2}
\title{Creates all trees for distance}
\usage{
allTrees2(hc, dis, start = 1)
}
\arguments{
\item{hc}{Hierarchical clustering object.}

\item{dist}{Distance object.}
}
\value{
Vector with markers with unclear relation (equal distances).
}
\description{
Takes order and tree and creates all possible trees
}
\keyword{internal}

