% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{calcSarfDist}
\alias{calcSarfDist}
\title{Calculates SARF}
\usage{
calcSarfDist(ord, dis, n = 1)
}
\arguments{
\item{ord}{Order of markers.}

\item{dis}{Distance object.}

\item{n}{Number of neighbors to include.}
}
\value{
SARF score.
}
\description{
Internal method to calculate the SARF score for a given order and distance object.
}
\keyword{internal}

