% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{findAmb}
\alias{findAmb}
\title{Finds ambiguous orders}
\usage{
findAmb(hc, dis)
}
\arguments{
\item{hc}{Hierarchical clustering object.}

\item{dist}{Distance object.}
}
\value{
Vector with markers with unclear relation (equal distances).
}
\description{
Takes order and tree and checks for ambiguity.
}
\keyword{internal}

