% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{findChr}
\alias{findChr}
\title{Find most equal chromosome in other map}
\usage{
findChr(map, comp)
}
\arguments{
\item{map}{Map to switch.}

\item{comp}{Other map for comparison.}
}
\value{
map
}
\description{
Find most equal chromosome in other map
}
\keyword{internal}

