% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{getIndependent_subtrees}
\alias{getIndependent_subtrees}
\title{Independent subtrees from tree}
\usage{
getIndependent_subtrees(merge, subtrees)
}
\arguments{
\item{merge}{Tree from hierarchical clustering.}

\item{subtrees}{List of leaf indices.}
}
\value{
Leaf indices, which are not in each others subtrees = independent.
}
\description{
Provides the independent subtree from a given tree and node.
}
\keyword{internal}

