% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{getLeaves}
\alias{getLeaves}
\title{Leaves from subtree}
\usage{
getLeaves(merge, i)
}
\arguments{
\item{merge}{Tree from hierarchical clustering.}

\item{i}{Node number.}
}
\value{
Leave indices.
}
\description{
Provides all leaves for a given tree and node.
}
\keyword{internal}

