% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{getNode}
\alias{getNode}
\title{Switches leaves in hclust object}
\usage{
getNode(merge, leave, max)
}
\arguments{
\item{hc}{Hierarchical clustering object.}

\item{pair}{Vector of length 2.}
}
\value{
Updated hclust object.
}
\description{
Switches leaves in hclust object
}
\keyword{internal}

