% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{getSubtree}
\alias{getSubtree}
\title{Subtrees from tree}
\usage{
getSubtree(merge, i = NULL, leaves = FALSE)
}
\arguments{
\item{merge}{Tree from hierarchical clustering.}

\item{i}{Node number.}

\item{leaves}{Logical. If TRUE the leaves will be provided.}
}
\value{
Vector containing indices.
}
\description{
Provides the subtree from a given tree and node.
}
\keyword{internal}

