% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{makealtord}
\alias{makealtord}
\title{Creates vectors with highly distant neighbors}
\usage{
makealtord(n = 3)
}
\arguments{
\item{n}{Length of vector.}
}
\value{
Vector of length n.
}
\description{
Creates a vector of numbers 1 to n, where the neighbors are as distant as possible.
For the grey scale plot, that guarantees, that the shades of grey are easy to distinguish.
For instance, 1, 4, 2, 5, 3; all numbers have a distance of at least 2 and where possible 3.
}
\keyword{internal}

