% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{map2dend}
\alias{map2dend}
\title{Transforming a map into a dendrogram}
\usage{
map2dend(map, method = "complete")
}
\arguments{
\item{map}{One map. Required.}
}
\value{
Dendogram object.
}
\description{
Create dendrogram object. The map specific distance are ignored and only the grouping and ordering is maintained.
Allows for comparison of whole map with package 'dendextend'
}
\examples{
data(simTetra)
simTetrageno<-bases2genotypes(simTetra,4)
rfMat<-calcRec(simTetrageno,4)
split<-splitChr(rfMat,nchr=7)
order<-sortLeafs(rfMat,split)
map<-pullMap(rfMat,order[split==1],names=rownames(simTetra)[split==1]) 
dend<-map2dend(map)  
plot(dend)
}

