% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recombination.R
\name{pairwRF}
\alias{pairwRF}
\title{Pairwise recombination frequency calculation}
\usage{
pairwRF(input, ploidy = 4, na.penalty = 0)
}
\arguments{
\item{input}{Matrix with two rows of genotypes. Each row is one marker.}

\item{ploidy}{Ploidy level of the organism.}
}
\value{
Numeric value of recomination between the two markers.
}
\description{
Calculates the pairwise recombination frequencies for two given markers.
Heuristic approach, assuming minimal recombination between markers.
}
\keyword{internal}

