% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola_package.R
\docType{package}
\name{pergola}
\alias{pergola}
\alias{pergola-package}
\title{Toolbox for polyploid genetic data}
\description{
This package provides multiple tools to work with polyploid data.
}
\details{
Load the dataset \code{simBC} and analyse it with \code{simpalyze}.
}

