% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{polyploid.hclust}
\alias{polyploid.hclust}
\title{Wrapping function for hierarchical clustering}
\usage{
polyploid.hclust(rfMat)
}
\arguments{
\item{rfMat}{Matrix of pairwise recombination frequencies.}
}
\value{
None.
}
\description{
Clusters with fixed parameters
}
\examples{
data(simTetra)
simTetrageno<-bases2genotypes(simTetra,4)
rfMat<-calcRec(simTetrageno,4)
polyploid.hclust(rfMat)
}
\keyword{internal}

