% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyploid.R
\docType{package}
\name{polyploid}
\alias{polyploid}
\alias{polyploid-package}
\title{Toolbox for polyploid genetic data}
\description{
This package provides multiple tools to work with polyploid data.
}
\details{
Load the dataset \code{simBC} and analyse it with \code{simpalyze}.
}

