% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{returnDup}
\alias{returnDup}
\title{Reinsert duplicates}
\usage{
returnDup()
}
\arguments{
\item{rf}{Matrix of pairwise recombination frequencies.}

\item{ord}{Vector with marker order.}

\item{split}{Vector of cluster numbers, created by splitChr(). Zeros indicated filtered markers and will be ignored.}
}
\value{
Vector with unique marker order.

Matrix with equally good marker orders.
}
\description{
Adds previously filtered duplicates back into the map at the same position as their duplicate.
}
\examples{
data(simTetra)
simTetrageno<-bases2genotypes(simTetra,4)
rfMat<-calcRec(simTetrageno,4)
split<-splitChr(rfMat,nchr=7)
order<-sortLeafs(rfMat,split)
oiu(rfMat,ord,split,max.n=6)   
}

