% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{sarfExt}
\alias{sarfExt}
\title{Extends SARF criterion to neighborhood}
\usage{
sarfExt(uniOrd, dis, maxSarf = length(uniOrd[1, ]) - 1)
}
\arguments{
\item{uniOrd}{Order of markers.}

\item{dis}{Distance object.}

\item{maxSarf}{Maximal number of neighbors to include. Should not exceed number of markers -1.}
}
\value{
Order with minimal SARF.
}
\description{
The neighborhood is increased stepwise until only one order remains.
}
\keyword{internal}

