% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simHexa}
\alias{simHexa}
\title{Hexaploid F2 population}
\format{A data frame with 131 rows and 600 variables:
\describe{
  \item{F2_001_1}{First allele in first offspring}
  \item{F2_001_2}{Second allele in first offspring}
  ...
  \item{F2_100_6}{Sixth allele in last offspring}
}}
\source{
\url{https://github.com/PBR/pedigreeSim}
}
\usage{
simHexa
}
\description{
100 offspring genotypes from an F2 crossing population.
Generated with PedigreeSim (Voorips et al, 2012).
}
\keyword{datasets}

