% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{splitChr}
\alias{splitChr}
\title{Split markers into chromosomes}
\usage{
splitChr(rf, height = 0.4, nchr = NULL, plotonly = FALSE,
  plottype = "dendrogram", method = "single", filter = FALSE,
  thresh = 0.05, cex.axis = 1, rm.dup = TRUE, ...)
}
\arguments{
\item{rf}{Matrix of pairwise recombination frequencies.}

\item{height}{Threshold value for grouping the markers.}

\item{nchr}{Expected number of chromosomes.}

\item{plotonly}{In case of unknown number of chromosomes, plotting the clustering tree, might give a hint.}

\item{plottype}{Default is "dendrogram". Any other value will plot the recombination frequencies.}

\item{method}{Default is "single", which is used for the hierarchical clustering.}

\item{filter}{Logical, if the result should be filtered or not. Default is FALSE. Creates zeros for the markers below the threshold.}

\item{thresh}{Threshold for filtering. Default is 0.05, i.e. linkage groups with less than 5\% of markers, are filtered out.}

\item{cex.axis}{Size of axis labels in image plot.}
}
\value{
Vector of cluster relationship. Same length and order as the matrix of recombination frequencies.
}
\description{
This function splits markers into linkage groups (LG), which ideally represent chromosomes.
The split is based on hierarchical clustering with a single linkage distance.
}
\examples{
data(simTetra)
simTetrageno<-bases2genotypes(simTetra,4)
rfMat<-calcRec(simTetrageno,4)
splitChr(rfMat,nchr=7)
###
splitChr(rfMat,nchr=7,plotonly=TRUE)
splitChr(rfMat,nchr=7,plotonly=TRUE, plottype="image")
}

