% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{swapChr}
\alias{swapChr}
\title{Swap chromosomes}
\usage{
swapChr(map, comp)
}
\arguments{
\item{map}{Map to switch.}

\item{comp}{Other map for comparison.}
}
\value{
map
}
\description{
Finds best matching chromosome for each chromosome and brings them into the same order.
}

