% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{switchChr}
\alias{switchChr}
\title{Switch Chromosomes}
\usage{
switchChr(map, comp)
}
\arguments{
\item{map}{Map to switch.}

\item{comp}{Other map for comparison.}
}
\value{
map
}
\description{
Switches chromosomes if comparison map fits better.
This does improve the comparability without changing the order, 
as chromosomes are independent.
}
\keyword{internal}

