% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{switchEntries}
\alias{switchEntries}
\title{Switches entries in hclust object}
\usage{
switchEntries(hc, e1, e2)
}
\arguments{
\item{hc}{Hierarchical clustering object.}

\item{pair}{Vector of length 2.}
}
\value{
Updated hclust object.
}
\description{
Switches entries in hclust object
}
\keyword{internal}

