% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{uniqueOrders}
\alias{uniqueOrders}
\title{Unique order}
\usage{
uniqueOrders(hc, dist, maxSarf = 4, maxPermut = 8)
}
\arguments{
\item{hc}{Hierarchical clustering object.}

\item{dist}{Distance object.}

\item{maxSarf}{Maximum neighborhood.}

\item{maxPermut}{Maximum number of markers in a subtree. Essential to the runtime of the function.}
}
\value{
Matrix of unique orders.
}
\description{
Takes order and tree and checks for ambiguity.
Creates all possible orders and puts them into a matrix.
}
\keyword{internal}

