% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{getNode}
\alias{getNode}
\title{Get Node for leave}
\usage{
getNode(merge, leave, max)
}
\arguments{
\item{hc}{Hierarchical clustering object.}

\item{pair}{Vector of length 2.}
}
\value{
Updated hclust object.
}
\description{
Finds the highest node of a leaf in hclust object below a threshold
}
\keyword{internal}

