% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pergola.R
\name{plotChr}
\alias{plotChr}
\title{Plotting one or two linkage maps}
\usage{
plotChr(map1, map2 = NULL, cex = 1, ...)
}
\arguments{
\item{map1}{One map. Required.}

\item{map2}{Optional second map for comparison.}

\item{cex}{Font size in the figure.}
}
\value{
None. Plotting only.
}
\description{
Visualization of one or two linkage maps.
Used as comparion between two different maps (e.g. different parameters or linkage mapping tools).
}
\examples{
data(simTetra)
simTetrageno<-bases2genotypes(simTetra,4)
rfMat<-calcRec(simTetrageno,4)
split<-splitChr(rfMat,nchr=7)
order<-sortLeafs(rfMat,split)
map<-pullMap(rfMat,order[split==1],names=rownames(simTetra)[split==1]) 
plotChr(map)  
}

