% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simTetra}
\alias{simTetra}
\title{Tetraploid F2 population}
\format{A data frame with 131 rows and 400 variables:}
\source{
\url{https://github.com/PBR/pedigreeSim}
}
\usage{
simTetra
}
\description{
100 offspring genotypes from an F2 crossing population.
Generated with PedigreeSim (Voorips et al, 2012).
}
\keyword{datasets}

