% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{switchChrs}
\alias{switchChrs}
\title{Switch Chromosomes}
\usage{
switchChrs(map, comp)
}
\arguments{
\item{map}{Map to switch.}

\item{comp}{Other map for comparison.}
}
\value{
map
}
\description{
Wrapper function to switch chromosomes for the whole map
}

